delete SelfUserInfoResponse.kt
