--- network/src/commonMain/kotlin/com/wire/kalium/network/api/user/self/SelfUserInfoResponse.kt
@@ -1,56 +1,0 @@
-package com.wire.kalium.network.api.user.self
-
-import com.wire.kalium.network.api.UserId
-import com.wire.kalium.network.api.UserSsoId
-import com.wire.kalium.network.api.model.ServiceDTO
-import com.wire.kalium.network.api.model.UserAssetDTO
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-// TODO: rename to UserDTO
-data class SelfUserInfoResponse(
-    @SerialName("email")
-    val email: String?,
-    @SerialName("phone")
-    val phone: String?,
-    @SerialName("sso_id")
-    val userSsoId: UserSsoId?,
-    @SerialName("id")
-    val id: String,
-    @SerialName("qualified_id")
-    val qualifiedId: UserId,
-    @SerialName("name")
-    val name: String,
-    @SerialName("accent_id")
-    val accentId: Int,
-    @SerialName("deleted")
-    val deleted: Boolean?,
-    @SerialName("assets")
-    val assets: List<UserAssetDTO>,
-    @SerialName("locale")
-    val locale: String,
-    @SerialName("service")
-    val service: ServiceDTO?,
-    @SerialName("expires_at")
-    val expiresAt: String?, // Time of sending message. ,
-    @SerialName("handle")
-    val handle: String?,
-    @SerialName("team")
-    val team: String?,
-    @SerialName("managed_by")
-    val managedBy: ManagedBy // 'wire', 'scim'
-)
-
-@Serializable
-enum class ManagedBy {
-    @SerialName("wire")
-    Wire,
-
-    @SerialName("scim")
-    Scim;
-
-    override fun toString(): String {
-        return this.name.lowercase()
-    }
-}
